ShoppingWebParent/ShoppingBackEnd/src/main/java/com/shopping/admin/brand/repository/BrandRepository.java
package com.shopping.admin.brand.repository;

import com.shopping.common.entity.Brand;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;
import org.springframework.data.repository.PagingAndSortingRepository;

public interface BrandRepository extends CrudRepository<Brand, Integer>,
        PagingAndSortingRepository<Brand, Integer> {

    @Query("SELECT b FROM Brand b WHERE b.name LIKE %?1%")
    public Page<Brand> findAll(String keyword, Pageable pageable);

    public Brand findByName(String name);

    Long countById(Integer id);
}
